refactor: Rename package 'best' to 'one2one'
