+++ bdikit/schema_matching/one2one/contrastivelearning.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from bdikit.schema_matching.best.base import BaseSchemaMatcher
+from bdikit.schema_matching.one2one.base import BaseSchemaMatcher
 from bdikit.models.contrastive_learning.cl_api import DEFAULT_CL_MODEL
 from bdikit.schema_matching.topk.contrastivelearning import CLTopkSchemaMatcher
