+++ docs/source/schema-matching.rst
@@ -16,13 +16,13 @@ To see how to use these methods, please refer to the documentation of :py:func:`
       - Class
       - Description
     * - ``ct_learning``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.ContrastiveLearningSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.contrastivelearning.ContrastiveLearningSchemaMatcher`
       - | Uses a contrastive (CT) learning model to learn embeddings for columns and retrieves the best match most similar columns using the cosine similarity between the column embeddings.
     * - ``two_phase``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.TwoPhaseSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.twophase.TwoPhaseSchemaMatcher`
       - | The two-phase schema matching method first uses a a top-k column matcher (e.g., `ct_learning`) to prune the search space (keeping only the top-k most likely matches), and then uses another column matcher to choose the best match from the pruned search space.
     * - ``gpt``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.GPTSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.gpt.GPTSchemaMatcher`
       - | This method uses the `ct_learning` to prune the search space and then uses a large language model (GPT4) to choose the best column match, given a set of top-k most likely candidates retrieved using the `ct_learning` method in the first phase.
 
 .. list-table:: Methods from other libraries
@@ -32,17 +32,17 @@ To see how to use these methods, please refer to the documentation of :py:func:`
       - Class
       - Description
     * - ``similarity_flooding``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.SimFloodSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.valentine.SimFloodSchemaMatcher`
       - | Similarity Flooding transforms schemas into directed graphs and merges them into a propagation graph. The algorithm iteratively propagates similarity scores to neighboring nodes until convergence. This algorithm was proposed by Sergey Melnik, Hector Garcia-Molina, and Erhard Rahm in "Similarity Flooding: A Versatile Graph Matching Algorithm and Its Application to Schema Matching" (ICDE, 2002).
     * - ``coma``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.ComaSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.valentine.ComaSchemaMatcher`
       - | COMA is a matcher that combines multiple schema-based matchers, representing schemas as rooted directed acyclic graphs. This algorithm was proposed by Do, Hong-Hai, and Erhard Rahm in "COMA — a system for flexible combination of schema matching approaches." (VLDB 2002). *This algorithm requires Java to be installed on the system.*
     * - ``cupid``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.CupidSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.valentine.CupidSchemaMatcher`
       - | Cupid is a schema-based approach that translates schemas into tree structures. It calculates overall similarity using linguistic and structural similarities, with tree transformations helping to compute context-based similarity. This algorithm was proposed by Madhavan et al. in "Generic Schema Matching with Cupid" (VLDB, 2001)​.
     * - ``distribution_based``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.DistributionBasedSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.valentine.DistributionBasedSchemaMatcher`
       - | Distribution-based Matching compares the distribution of data values in columns using the Earth Mover's Distance. It clusters relational attributes based on these comparisons. This algorithm was proposed by Zhang et al. in "Automatic discovery of attributes in relational databases" (SIGMOD 2011).
     * - ``jaccard_distance``
-      - :class:`~bdikit.mapping_algorithms.column_mapping.algorithms.JaccardSchemaMatcher`
+      - :class:`~bdikit.schema_matching.one2one.valentine.JaccardSchemaMatcher`
       - | This algorithm computes pairwise column similarities using Jaccard similarity, treating values as identical if their Levenshtein distance is below a threshold. The algorithm was proposed by Koutras et al. in "Valentine: Evaluating matching techniques for dataset discovery" (ICDE 2021).
